requirements.txt
selenium==4.10.0
beautifulsoup4==4.12.2
pandas==1.3.0
loguru==0.7.0
minio==7.1.16
PyYAML==6.0.1
deltalake==0.10.2
numpy==1.21.4
pyspark==3.5.0
psycopg2-binary==2.9.7
sqlalchemy==1.4.49
confluent-kafka==2.2.0
python-dotenv==1.0.0
apache-flink==1.17.1
catboost==1.2.1
fastapi==0.96.0
joblib==1.1.0
loguru==0.7.0
uvicorn[standard]==0.22.0
apache-airflow==2.7.1
apache-airflow-providers-postgres==5.6.1
airflow-provider-great-expectations==0.2.6
apache-airflow-providers-docker==3.7.5
great_expectations==0.17.21
pydeequ==1.0.1
joblib==1.1.0
# ansible==8.3.0
kafka-python==2.0.2